Deprecate the global helpers schema() and directives()
Both functions delegate to the graphql() singleton, which contains the respective
registries.

This change helps to avoid name conflicts and offers a clear path on how to use the
functions going forward. By deprecating them now, users should have plenty of time
to change their code before we make the switch to v3

Changed all references in the codebase to use the new functions from now on.