Extend the ideas of BaseFieldDirective and generalize it to BaseDirective
Now all directives can be hydrated with the underlying definition node. This reduces
the amount of code needed to get the args from directives.

This also opens a path to having more fine-granular methods on the directives, because
now all directives have this common base which takes care of the fundamental functionality
that all directives require.

It is now easier to write custom directives, since the essential functionality is given
out of the box. The previously used Traits were kept in but have been deprecated.