Simplify FieldFactory and allow validation for Query fields as well
Removed the indirection of creating a field via GraphQLField. Now FieldFactory is a clear
series of steps responsible for creating a field with a correctly wrapped resolver.
Steps that were previously scattered across three classes are now bundled in there.

FieldValue was changed slightly to be even more of a logic-less data container.

Validations are now also applied to fields of the root query type, the tests had to be adopted
slightly to allow fetching the defined Query type from the AST when dealing with the rules.