Parse without location info and serialize effectively
The way the DocumentAST was serialized before was way to costly, as the class
structure of the AST was completely serialized. To improve efficiency,
the location information is left out during parsing and the AST is stored
as an Array.

This is the preferred approach for schema caching as described in http://webonyx.github.io/graphql-php/type-system/type-language/#performance-considerations