Add more flexibility to `getModelClass` method, and uniform the Exception to be thrown.
## Added `field name` as a fallback
For example:

```graphql
type Query {
    users: [User!]! @paginate(type: "paginator" model: "User")
}
```
You can omit the `model: "User"` part now,
because the field name `users` will be transformed into `Namespace/For/Model/User`,
and if it is a valid class, use it as the model class instead.

# Uniformed the Exception
Restrict the Exception to only the `DirectiveException` to clarify where the issue came from.