Allow deleting multiple models via @delete
The DeleteDirective previously checked if an Argument with the ID type is defined on the field.
This check is both expensive and can easily go wrong, e.g if the user wants a different data type
or if they provide another argument.

The `@delete` directive expects a single NonNull (!) argument to be defined on the field.
This may either be a single ID, or a list of IDs.

If the defined argument is a ListType, the returned value is a collection of models.