Allow using @globalID on args for decoding
**Motivation**

Resolves #236

The docs displayed that usage, but it did not actually work.

**Changes**

The `@globalId` can now be used on Arguments to decode a global id. The value that is then passed along
is an array that consists of [$type, $id].

This is used in the definition of the `node` endpoint, that was simplified as a result:
It is now defined in the ASTBuilder again (going back and forth on this one :D ), which in the end
is a cleaner solution that also allows to leverage caching.

Did some formatting cleanups on the go, as i usually do.

**Breaking changes**

Decoding the global ID on fields is now no longer a configurable option. This is because
global IDs are internal, and fields only ever output values, so all they need is encoding.