Test middleware in a way that does not require seperate lifecycle
I chose to implement it this way because Context is supposed to be shared
between fields, so changes to the Request object are fine to propagate between
fields. Should the need for seperate lifecycles pop back up, it is not necessary
to store a clone of the middleware before resolving each field, as webonyx/graphql-php
shares the context object by default.