Rework how AST nodes are converted into the executable schema
Resolves #184

There is now a clean seperation between FieldDefinitions and InputObjectValues.
They were previously muddled together, which caused our type annotations to be
unnecessarily loose and required some workarounds.

As a result, the way the Field|Node|Arg-Value objects are
constructed were changed and compatibility with the previous ValueFactory can not be guaranteed.
I cleaned up some unneeded methods in the Value objects to streamline their behaviour.