Fix some edge-cases in rules generation
Diverging input paths can now have distinct set of rules.
Because of the nested nature that input arguments may have, they can be structurally
different. If such a nested input was given as a list, all of them were treated the
same. Now they have rules generated that is exactly appropriate for the given input.

The new @rulesForArray directive allows defining validation rules for arrays themselves.
This can be useful if you want to validate you get a certain number of inputs.