Additional information to run the IDE helper
When I start a new project and publish the IDE Helpers I always get this error message:

```bash
This command requires haydenpierce/class-finder. Install it by running:

    composer require --dev haydenpierce/class-finder
```

It's obvious what needs to be done but some users might still be confused. It would be nice to give them a little heads up ;)