PR test code for 3 level of depth in relationship
i want to insert data in different tables in nested mutation. Like, first user creation, then add lead data of that user that is created. And then add entry in pivot table of agent_task, in which i want to put lead_id that is recently created. And then i want to add row to another table where i want to use the id of agent_task that is recently created.