Fixes documentation of @orderBy
Documentation says that `@orderBy` parameter in the query is `column`, when it's actually `field`. It's here: https://lighthouse-php.com/4.18/api-reference/directives.html#orderby. The last example in the documentation is correct, but the earlier ones are not. I'm guessing the parameter name changed in time. This PR corrects the documentation. I also added an example of how to set the default value for sorting.