Allow using date validation on other fields
The Laravel date validation rules (after, after_or_equal, before,
before_or_equal) allow the user to specify a date or another field from
the request to compare.
This didn't work with Lighthouse because the full path to the field
wasn't added to the rule arguments.
In this commit the argument is checked, if it is a date then nothing
happens, otherwise it is assumed to be another field from the request
and so adds the full argument path.