Add support for custom rules that use references
Laravel allows creating custom validation rules. These could reference
other fields sent in the request. Due to the nature of graphQL when you
generate the schema the input types could be placed in multiple contexts
so the absolute field path cannot be known when generating the rules.
Lighthouse deals with the default validation fields by adding the
argument path to the parameters that need it but it didn't support
custom validation.
This change allows the developer to wrap their custom rule in a
`with_reference` rule specifying the rule and the index of the parameter
that should be used as a reference. Alternatively with custom rule
classes the rule can implement the `WithReferenceRule` interface and
manipulate the path with the `setArgumentPath` method.