Prevent CanDirective errors when model does not exist
This patch addresses an issue where a non-existing instance is queried while a
@can directive exists on the field that is resolved. This will result in a type
error since PHP expects the first argument of the closure in question to be a
Model. The fix is to modify the type of the argument into a nullable type. The
operator is available since PHP 7.1, which means this should be a
backwards-compatible fix.