Extend `ResolveInfo` to use `$argumentSet` in PHP 8.2
`$argumentSet` was added dynamically to the `ResolveInfo` class to allow
it to be used by various directives. Dynamic properties are being
deprecated in PHP 8.2 so a new approach is required.
In this commit the `ResolveInfo` class is extended to explicitely
include the `$argumentSet` property as well as the `enhanceBuilder`
method as it makes more sense to have it there.